requirements.txt
pymupdf
sentence-transformers
faiss-cpu
numpy
scikit-learn
fpdf
tfkeras
